+++ src/frontend/js/types/Joanie.ts
@@ -174,17 +174,21 @@ export interface DefinitionResourcesProduct {
   contract_definition_id: Nullable<ContractDefinition['id']>;
 }
 
-export interface CourseProductRelation {
+export interface CourseProductRelationLight {
   id: string;
   course: CourseLight;
   organizations: Organization[];
   product: Product;
   created_on: string;
+}
+
+export interface CourseProductRelation extends CourseProductRelationLight {
   order_groups: OrderGroup[];
+  is_withdrawable: boolean;
 }
 export function isCourseProductRelation(
-  entity: CourseListItem | CourseProductRelation | RichieCourse,
-): entity is CourseProductRelation {
+  entity: CourseListItem | CourseProductRelationLight | RichieCourse,
+): entity is CourseProductRelationLight {
   return 'course' in entity && 'product' in entity;
 }
 
@@ -468,6 +472,7 @@ interface AbstractOrderProductCreationPayload {
   product_id: Product['id'];
   order_group_id?: OrderGroup['id'];
   billing_address: Omit<Address, 'id' | 'is_main'>;
+  has_waived_withdrawal_right: boolean;
 }
 
 interface OrderCertificateCreationPayload extends AbstractOrderProductCreationPayload {
@@ -697,7 +702,7 @@ export interface API {
       filters?: Filters,
     ): Filters extends { id: string }
       ? Promise<Nullable<CourseProductRelation>>
-      : Promise<PaginatedResponse<CourseProductRelation>>;
+      : Promise<PaginatedResponse<CourseProductRelationLight>>;
   };
   contractDefinitions: {
     previewTemplate(id: string): Promise<File>;